+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+- Changed BasicFileSink to use `print` instead of `stdout.write`
+- Added UniversalIo for better web interopoerability
+
 ## 0.1.0
 
 - Initial version.